move App interface to simapp package
